refactor:공고상세조회 / 필터 시트 역할,책임 분리 작업
