+++ src/features/listings/ui/listingsCardDetail/components/DetailSectionFilter/DistanceFilter.tsx
@@ -9,34 +9,22 @@ import {
   mapPinPointToOptions,
 } from "@/src/features/listings/hooks/listingsHooks";
 import { ListingCardDetailOut } from "@/src/features/listings/ui/listingsCardDetail/button/button";
+import {
+  useDistanceHooks,
+  useDistanceVariable,
+} from "@/src/features/listings/ui/listingsCardDetail/hooks/distanceHooks";
 
 const SLIDER_MIN = 0;
 const SLIDER_MAX = 120;
 
 export const DistanceFilter = () => {
   const { data, isFetching } = useListingFilterDetail<PinPointPlace>();
-  const pinPointData = data?.pinPoints;
-  const pinPointList = mapPinPointToOptions(pinPointData);
-  const dropDownTriggerLabel = getDefaultPinPointLabel(pinPointList);
-  const hasPinPoints = pinPointList.myPinPoint.length > 0;
-  const { setPinPointId } = useOAuthStore();
-  const { distance, setDistance } = useListingDetailFilter();
-
-
-  const onChageValue = (selectedKey: string) => {
-    setPinPointId(selectedKey);
-  };
-
-  const handleDistanceChange = (values: number[]) => {
-    const [nextValue] = values;
-    if (typeof nextValue === "number") {
-      setDistance(nextValue);
-    }
-  };
-
-  const sliderValue = [distance];
-  const formatMinutes = (value: number) => value.toString().padStart(1, "0");
-  const formattedDistance = formatMinutes(distance);
+  const emptyPinPoint: PinPointPlace = { userName: "", pinPoints: [] };
+  const { pinPointList, dropDownTriggerLabel, hasPinPoints } = useDistanceVariable(
+    data ?? emptyPinPoint
+  );
+  const { onChangeValue, handleDistanceChange, sliderValue, formattedDistance } =
+    useDistanceHooks();
 
   return (
     <div className="flex h-full flex-col">
@@ -54,7 +42,7 @@ export const DistanceFilter = () => {
           types="myPinPoint"
           data={pinPointList}
           size="lg"
-          onChange={onChageValue}
+          onChange={onChangeValue}
           disabled={isFetching || !hasPinPoints}
         >
           {dropDownTriggerLabel}
@@ -75,7 +63,6 @@ export const DistanceFilter = () => {
           labelSuffix="분"
         />
       </section>
-
     </div>
   );
 };
